Quick fix in onedspec, making the number of components an integer.

+++ ifscube/onedspec.py
@@ -298,7 +298,7 @@ class Spectrum():
 
         if component_names is None:
             self.component_names = [
-                'C_{:03d}'.format(i) for i in range(npars / npars_pc)]
+                'C_{:03d}'.format(i) for i in range(int(npars / npars_pc))]
         else:
             self.component_names = component_names
